+++ tests/minimal.lua
@@ -33,16 +33,16 @@ if vim.fn.empty(vim.fn.glob(packer_install_path)) > 0 then
   install_plugins = true
 else
   vim.cmd('packadd packer.nvim')
-  vim.cmd('packadd plenary.nvim')
-  vim.cmd('runtime! plugin/plenary.vim')
 end
 
 local packer = require('packer')
 
 packer.init {
-  package_root = data_path .. '/pack',
+  package_root = data_path .. '/site/pack',
   compile_path = data_path .. '/plugin/packer_compiled.lua',
 }
+
+vim.cmd('runtime! plugin/plenary.vim')
 
 packer.startup(function(use)
   use('wbthomason/packer.nvim')