+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for `hls-cabal-plugin` and `hls-cabal-fmt` plugins.
 
 ## [1.4.4] - 2022-12-20
 ### Fixed