+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   themes, etc. can be supported.
 
 ## [1.4.3] - 2022-12-06
+### Added
+- rockspec for automatic dependency management by luarocks
 ### Fixed
 - Error message shown if hoogle is installed, but telescope is missing